requirements.txt
opencv-python
tensorflow-gpu==2.5.0
tqdm